[FEATURE] Pattern-match ref objects
This is an unholy abomination that was added in order to be somewhat
more consistent wrt to `of` operator, but it works quite well though.
And didn't take that long to implement (most of the changes in this
commit come from `$` operators for `Match` and related things). All
thanks to MCS